Fix 32-bit failure

+++ test/runtests.jl
@@ -762,7 +762,7 @@ end
     @test quantile(Float32[-1e15, -1e14, -1e13, -1e12, -1e11, -1e10, -1e9, 100], 1) ≈ 100
     @test quantile(Int8[-68, 60], 0.5) ≈ -4
     @test quantile(Int32[-1e9, 2e9], 1.0) ≈ 2.0e9
-    @test quantile(Int[-5e18, -2e18, 9e18], 1.0) ≈ 9.0e18
+    @test quantile(Int64[-5e18, -2e18, 9e18], 1.0) ≈ 9.0e18
 end
 
 @testset "variance of complex arrays (#13309)" begin
